changed back the return type of MortgageCalculator.calculateMortgagePayment()

+++ app/src/main/java/dev/andrylat/raqimbek/bankingutils/MortgageCalculator.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 import java.math.RoundingMode;
 
 public class MortgageCalculator {
-  public long calculateMortgagePayment(BigDecimal p, double r, int n, int t) {
+  public double calculateMortgagePayment(BigDecimal p, double r, int n, int t) {
     // formula: (p * (r / n) * (1 + ((r / n) ^ n) * t)) / ((1 + ((r / n) ^ n) * t) - 1)
 
     var annualInterestRate = new BigDecimal(r / n);
@@ -13,6 +13,6 @@ public class MortgageCalculator {
     return p.multiply(annualInterestRate)
         .multiply(bg1)
         .divide(bg1.subtract(BigDecimal.ONE), 10, RoundingMode.HALF_EVEN)
-        .longValue();
+        .doubleValue();
   }
 }
